Set diversity groups the same way as query labels
There should be fewer bugs if we remove the code that tries to ignore
one of the columns of the main `X` array.

+++ include/LightGBM/dataset.h
@@ -98,7 +98,7 @@ class Metadata {
   */
   void SetInitScore(const double* init_score, data_size_t len);
 
-  void SetDiversityGroups(const double* group_labels, data_size_t len);
+  void SetDiversityGroups(const int* group_labels, data_size_t len);
 
 
   /*!
